♻️[Refactor]#7 - 쪽지 DTO 파일 분리
