+++ Muje/Muje/Models/DTO/Conversation.swift
@@ -109,19 +109,3 @@ extension Conversation: EntityRepresentable {
     }
 }
 
-struct Message: Identifiable, Codable, Equatable {
-    @DocumentID var id: String?
-    let senderUserId: String
-    let text: String
-    @ServerTimestamp var createdAt: Timestamp?
-    @ServerTimestamp var updatedAt: Timestamp?
-
-    enum CodingKeys: String, CodingKey {
-        case senderUserId = "sender_user_id"
-        case text
-        case createdAt = "created_at"
-        case updatedAt = "updated_at"
-    }
-
-    var createdDate: Date { createdAt?.dateValue() ?? .init() }
-}
